Downgrading OpenMPI from v5.x to v4.1.4
This downgrade is essentially to accomodate the SLURM's MPI's
process management interface-2 (PMI-2). If one enters
```
arc$ srun --mpi=list
```
It can be viewed that pmi2 is the only interface provided. OpenMPI
v5.x unforuntately dropped the support for pmi2 and only supports pmix.

Therefore, I have downgraded the ARC's OpenMPI to v4.1.4.

Furthermore, the path of SLURM's library and headers has been changed
to a project space's path. The details will be instructed in a future commit.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>